Assignment 3 - Mutation Testing

See merge request cse2115/2022-2023/SEM22a!43

+++ hoa-microservice/src/main/java/nl/tudelft/sem/template/hoa/models/ProposalRequestModel.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProposalRequestModel {
-    public int hoaId;
+    public long hoaId;
     public String name;
     public String description;
     public TimeModel scheduledFor;
